fix:bot profit issue
