+++ src/api/deposit.ts
@@ -2,8 +2,6 @@ import express from 'express';
 import { iStakeInfo, StakeInfo } from '../models/stakeInfoModel';
 import { User } from '../models/userModel';
 import { Bot, iBot } from '../models/botModel';
-import { Balance, iBalance } from "../models/balanceModel";
-import { getBalance } from '../services/stargateClient';
 
 const router = express.Router();
 
@@ -36,14 +34,6 @@ router.post('/api/deposit', async (req, res) => {
             amount,
         });
         await newStakeInfo.save();
-        
-        // const balance: iBalance | null = await Balance.findOne({ bot_id: bot_id }).sort({ timestamp: -1 }).exec();
-        
-        // if (!balance) {
-        //     return res.status(404).json({ success: false, message: 'Balance not found' });
-        // }
-        // balance.balance += amount;
-        // await balance.save();
 
         res.json({ success: true, balance: user.stakeAmount });
     } catch (error: any) {
