+++ src/server.ts
@@ -2,9 +2,7 @@ import app from './app';
 import fs from "fs";
 import https from "https";
 import { scheduleJob } from 'node-schedule';
-import { saveBotBalance } from './services/stargateClient';
-
-
+import { saveBotBalance } from './services/balanceService';
 
 const HTTP_PORT = process.env.HTTP_PORT || 3000;
 const HTTPS_PORT = process.env.HTTPS_PORT || 4000;
@@ -13,7 +11,7 @@ app.listen(HTTP_PORT, () => {
     console.log(`Server running on port ${HTTP_PORT}`);
 
     scheduleJob('0 0 * * *', async function () {
-        saveBotBalance();
+        await saveBotBalance();
     });
 });
